agents/alex/mark/requirements.txt
anthropic>=0.39.0
streamlit>=1.38.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
readability-lxml>=0.8.1
lxml>=5.0.0
pandas>=2.1.0
plotly>=5.18.0
html2text>=2024.2.26
